refactor [__init__] - settings: settings class on its own module
